% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-Params.R
\name{AssignMSMSParam}
\alias{AssignMSMSParam}
\title{AssignMSMSParam}
\usage{
AssignMSMSParam(
  methodAssign = c("highest"),
  methodRT = c("raw", "corrected"),
  rtTol = 0,
  rtExtend = 90,
  mzTol = 0.01,
  ccsTol = 0.5,
  rtUnitMS1 = c("s", "m"),
  rtUnitMS2 = c("s", "m"),
  isolationWindow = 1,
  thrIsolationIntensity = 150,
  thrIntMS1Predicted = 500,
  snthreshMSMS = 10,
  minfracVote = 0.25,
  outputSpec = FALSE,
  plotMSMSEIC = TRUE,
  typeMSMSEIC = c("pdf", "png", "both"),
  rerun = FALSE
)
}
\arguments{
\item{methodAssign}{\code{character} Method for MSMS asignment
\itemize{
  \item highest - select the spectrum with highest precursor intensity when
   fragmentation in all samples
  \item consensus - consensus of all found spectra
  \item combined" - consensus of spectra selected from each sample with
   highest precursor intensity when fragmentation
}}

\item{methodRT}{\code{character} Raw or corrected RT for MSMS assignment}

\item{rtTol}{\code{numeric} RT tolerance when matching MSMS spectra and MS1 peaks}

\item{rtExtend}{\code{numeric} Extended RT  tolerance  when matching MSMS
spectra and MS1 peaks}

\item{mzTol}{\code{numeric} MZ tolerance when matching MSMS spectra and MS1 peaks
\itemize{
  \item mzTol < 1 tolerance in Da
  \item mzTol >= 1 tolerance in ppm
}}

\item{ccsTol}{\code{numeric} CCS tolerance (in percentage) when matching MSMS
spectra and MS1 peaks}

\item{rtUnitMS1}{\code{character} RT unit for user provided MS1 peak table}

\item{rtUnitMS2}{\code{character} RT unit for user provided MSMS spectrum files}

\item{isolationWindow}{\code{numeric} Isolation window when selecting precursor
ions for fragmentation}

\item{thrIsolationIntensity}{\code{numeric} Intensity threshold of the
contaminat precorsor ions in isolation window to be extracted}

\item{thrIntMS1Predicted}{\code{numeric} Intensity threshold of precursor ions
for fragmentation to be kept}

\item{snthreshMSMS}{\code{numeric} S/N threshold for MSMS fragments}

\item{minfracVote}{\code{numeric} Minimal fraction for presentation of fragment
to be kept in consensus spectrum}

\item{outputSpec}{\code{logical} If output spectra to the spectral MSP file}

\item{plotMSMSEIC}{\code{logical} If plot EIC with found MSMS specta}

\item{typeMSMSEIC}{\code{character} Image format (either in PNG or PDF) for plotting}

\item{rerun}{\code{logical} If rerun all data processing progress}

\item{ppmBinMZ}{\code{numeric} Minimal mz in ppm for fragment binning}

\item{minBinMZ}{\code{numeric} Minimal mz in Da for fragment binning}

\item{thrRingMz}{\code{numeric} MZ tolerance in Da for finding satelite fragments}

\item{thrRingIntRel}{\code{numeric} Relative intensity tolerance for finding
satelite fragments}
}
\description{
Parameters for MSMS assignment
}
