% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function-Params.R
\name{ExperimentParam}
\alias{ExperimentParam}
\title{Experiment Setup}
\usage{
ExperimentParam(
  wd = ".",
  resDir = "results",
  tmpDir = file.path(resDir, "tmp"),
  polarity = c("positive", "negative"),
  ce = "30",
  lc = c("RP", "HILIC"),
  instrument = c("Sciex6600", "Agilent6560"),
  ms1range = NULL,
  ms2range = NULL,
  resDefineAt = 400,
  nSlaves = 4
)
}
\arguments{
\item{wd}{\code{character} Working directory for experiment to be processed}

\item{resDir}{\code{character} Folder for storing results}

\item{tmpDir}{\code{character} Folder for storing files when processing. By default,
it is the same with \code{wd}}

\item{polarity}{\code{character} Ion mode for MS data acquisition}

\item{lc}{\code{character} LC column used for LC seperation}

\item{instrument}{\code{character} Instrument used for MS data acquisition}

\item{ms1range}{\code{numeric(2)} Mass range setup for MS1 acquisition}

\item{ms2range}{\code{numeric(2)} Mass range setup for MS2 acquisition}

\item{resDefineAt}{\code{numeric} Mz value for defining resolution}

\item{nSlaves}{\code{numeric} Number of threads to be used}
}
\value{
an \code{ExperimentParam} object
}
\description{
Parameters for experimental setup
}
